fix: backend 코드 품질 CRITICAL/HIGH 14건 수정
code-analyzer 분석 결과 기반 보안·안정성 이슈 일괄 수정:

CRITICAL (5건):
- config.py: DATABASE_URL, JWT_SECRET 하드코딩 기본값 제거 → 필수 환경변수화
- database.py: Celery용 sync_session_maker 추가 (psycopg2 동기 세션)
- rbac.py: PERMISSIONS에 "startup" 리소스 추가
- fund.py: benchmark_return_rate float → Decimal
- contract.py: investment_amount, pre_money_valuation Integer → Numeric(20,2)

HIGH (8건):
- handover/review/contract/dashboard_service: datetime.now(timezone.utc) 통일
- incubation/lp/partner_demand_service: escape_like() 적용 (SQL 인젝션 방지)
- fund_service: get_by_id에 is_deleted 필터 추가
- lps.py: LP CUD에 require_permission("contract","full") RBAC 적용
- team_kpis.py: create/seed에 require_permission("kpi","full") RBAC 적용

Additional:
- follow_on_investment.py: target_amount Integer → BigInteger

Gap 분석 Match Rate: 100% (14/14)

+++ backend/app/models/contract.py
@@ -20,8 +20,8 @@ class InvestmentContract(Base):
     status: Mapped[ContractStatus] = mapped_column(default=ContractStatus.IC_RECEIVED)
 
     # 투자 조건
-    investment_amount: Mapped[int] = mapped_column(Integer)
-    pre_money_valuation: Mapped[int] = mapped_column(Integer)
+    investment_amount: Mapped[Decimal] = mapped_column(Numeric(20, 2))
+    pre_money_valuation: Mapped[Decimal] = mapped_column(Numeric(20, 2))
     equity_pct: Mapped[Decimal] = mapped_column(Numeric(10, 4))
     vehicle: Mapped[InvestmentVehicle] = mapped_column()
     follow_on_rights: Mapped[bool] = mapped_column(default=False)
